refactor: renomeia DataTable para CoeqDataTable
